+++ examples/interpret.rs
@@ -4,7 +4,7 @@ extern crate parity_wasm;
 
 use std::env::args;
 
-use parity_wasm::{interpreter, ModuleInstanceInterface};
+use parity_wasm::ModuleInstanceInterface;
 
 fn main() {
     let args: Vec<_> = args().collect();
